+++ src/components/party/PartySelect.tsx
@@ -1,11 +1,11 @@
-import { useListParties } from "../../api/tabletop-almanac/useParty.ts";
-import { Loader } from "../general/Loader.tsx";
 import { useMetadataContext } from "../../context/MetadataContext.ts";
 import { Link } from "@mui/material";
 import { useEffect, useState } from "react";
 import { useDebounce } from "ahooks";
 import { updateRoomMetadata } from "../../helper/helpers.ts";
 import { usePartyStore } from "../../context/PartyStore.tsx";
+import { useListParties } from "../../api/tabletop-almanac/useParty.ts";
+import { Loader } from "../general/Loader.tsx";
 
 export const PartySelect = () => {
     const room = useMetadataContext((state) => state.room);